Keep a stopped story's grapher and player pipeline for the whole delay
GrapherDataStore and PlayerDataStore multiplied the 32-bit
inactive_pipeline_delay_secs by 1000000000 before widening it, so any
delay of 5 s or more wrapped: the template's 300 s became 3.65 s. A native
run shows it: stop_story_recording at 14:53:35.7, retirement at the next
loop pass after 14:53:39.4.

On the grapher, retirement writes every window, empty ones too, so W jumped
to the end of the story's timeline about 4 s after release. The keepers'
last chunks, sealed 10 to 25 s after release, then always arrived as
orphans. The grapher adopted each one into a new pipeline and wrote it as
a second, numbered file for the same window.

The delay is now widened to 64 bits before the multiplication. The new
tests stop a story with a 5 s delay and check that its pipeline is still
there after 1.5 s, and gone after the delay when it is 1 s.

+++ src/chrono-grapher/src/GrapherDataStore.cpp
@@ -143,8 +143,9 @@ int chronolog::GrapherDataStore::stopStoryRecording(chronolog::StoryId const& st
     auto pipeline_iter = theMapOfStoryPipelines.find(story_id);
     if(pipeline_iter != theMapOfStoryPipelines.end())
     {
+        // widen first: the delay is 32-bit, and 5 s or more wraps in 32-bit nanoseconds
         uint64_t exit_time = std::chrono::high_resolution_clock::now().time_since_epoch().count() +
-                             inactive_pipeline_delay_secs * 1000000000;
+                             static_cast<uint64_t>(inactive_pipeline_delay_secs) * 1000000000ULL;
         // (*pipeline_iter).second->getAcceptanceWindow();
         pipelinesWaitingForExit[(*pipeline_iter).first] =
                 (std::pair<chl::StoryPipeline*, uint64_t>((*pipeline_iter).second, exit_time));
